+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31">
         <activity
-            android:name=".CartActivity"
+            android:name=".presentation.cart.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".ProductDetailActivity"
+            android:name=".presentation.productdetail.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ProductActivity"
+            android:name=".presentation.product.ProductActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
