+++ app/src/main/java/woowacourse/shopping/domain/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.domain
 
 data class Product(
     val name: String,
