+++ app/src/main/res/layout/activity_cart.xml
@@ -11,7 +11,7 @@
         android:id="@+id/cl_cart"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CartActivity">
+        tools:context=".presentation.cart.CartActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tb_cart"
